gui/build: create build script for detecting the exported haskell stub into a c++ executable to invoke the runtime from c++.
This is needed to control which  thread handles the GUI.
On MacOS, only the main thread is allowed to handle GUI events.
Having the program start through haskell removes our control of
where the executions happen.

Now, the execution begins in a c++ main and execution of the
haskell runtime can be delegated to another thread while the main
thread will be reserved for handling GUI events.